Moderization & Discord Bot * Mud <> Discord functionality * Update Dockerfile to use Ubuntu 20.04, clang 12 * Rework child_died to poll based * Add fmt library * Iterators --> for(auto) * Socklist from std::list<Socket*> --> std::list<Socket> *** Reboot doesn't fully work, either fix or disable *** * Make more things private * Fix bug loading longs with toNum * Add typedefs for <int, *blah> --> BlahMap

+++ include/xml.hpp
@@ -138,7 +138,7 @@ namespace xml {
             return(toReturn);
 
         try {
-            toReturn = static_cast<Type>(lexical_cast<int>(fromStr));
+            toReturn = static_cast<Type>(lexical_cast<long>(fromStr));
         } catch (bad_lexical_cast &) {
             // And do nothing
         }
